rename seven_zip.rb to better match class name
