+++ lib/cask/container.rb
@@ -11,7 +11,7 @@ require 'cask/container/gzip'
 require 'cask/container/naked'
 require 'cask/container/otf'
 require 'cask/container/pkg'
-require 'cask/container/sevenzip'
+require 'cask/container/seven_zip'
 require 'cask/container/sit'
 require 'cask/container/tar'
 require 'cask/container/ttf'
