+++ lib/cask/container/seven_zip.rb
